Set ALLOW_MEMORY_GROWTH emscripten build flag
This allows Hext to use more than the default amount of
memory allocated (8mb?), which, when reached without this
flag being set, causes Hext to explode with an exception.
Parsing a large HTML document will cause this. This flag
gets around this limitation and allows Hext to allocate
as much RAM as it normally would outside of emscripten.