Include JS/WASM assets in repo so we can use jsdelivr
Eventually we'll want to remove these since we already have
them being build in our releases and stored in the assets.
But jsdelivr still doesn't support asset-only releases
so the best way to use github as our CDN is to include these
in the repo.

Pending https://github.com/jsdelivr/jsdelivr/issues/18203
we will remove the assets from the repo.