Fully loaded hext.js ES6 module
This build uses emscripten's ES6 module settings, along with
a custom outer JS appended script, to provide a module that
can be loaded without the standard emscripten async WASM/loading
process. So now we have a `hext.js` that exports `hext` by default
which contains `Rule` and `Html` (which are also non-default exports).

The appended JS fragment, `wrapper/extern-post.js` contains almost
everything needed to fully encapsulate the exports we need,
with one exception: Emscripten doesn't let us disable the default
export of the loader function (`loadHext`) and also use the ES6 module
closure, so we remove it from the script using `sed` at the very end of
the Makefile. If we still find the default export in the `hext.js`, we
fail the build.

We might be able to do this via an emscripten build flag in the future,
but a similar issue [1] was brought up and the conclusion was a custom
flag isn't needed yet.

[1]: https://github.com/emscripten-core/emscripten/issues/13894